+++ Assets/Pedestrian/PedestrianPosition.cs
@@ -3,12 +3,12 @@ using System;
 public class PedestrianPosition
 {
 	private int id;
-	private decimal time;
+	private float time;
 	private float x;
 	private float y;
     private float z;
     private float zOffset = 0.3f;
-	public PedestrianPosition (int id, decimal time, float x, float y, float z) {
+	public PedestrianPosition (int id, float time, float x, float y, float z) {
 		this.id = id;
 		this.time = time;
 		this.x = x;
@@ -17,7 +17,7 @@ public class PedestrianPosition
 	}
 
 	public int getID() {return this.id;}
-	public decimal getTime() {return this.time;}
+	public float getTime() {return this.time;}
 	public float getX() {return this.x;}
 	public float getY() {return this.y;}
     public float getZ() {return this.z;}
